Windows application icon
For the windows platform the .ico file get embedded in the executable and shows up in file listing of the executable , in the task bar, window switch etc.

The png is file is probably needed for Linux.

The icon itself was modified from the logo on the language site with a round feature and inverted colors, in order to function visually.  This could probably be done better by utilizing the original file (SVG) and is a proposal.